# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Build Governance Platform",
...
    "version": "0.1.0-bgp-remediation"
  },
  "paths": {
...
    "/orgs": {
      "post": {
        "tags": [
          "orgs"
        ],
        "summary": "Create Org",
        "description": "REQ-001: explicit organisation membership before project access --\nthe creator becomes tenant_administrator of their own new tenant, not of\nany tenant they didn't create or weren't invited to.\n\nOrg bootstrap fix: `memberships` is RLS-protected, fail-closed with no\n`app.tenant_id` set -- there is no existing membership dependency to set\nit here (this IS the request that creates the first one), so this\nendpoint sets it itself, from the just-generated tenant.id, before the\nINSERT. Without this the INSERT was rejected outright\n(`new row violates row-level security policy`) against real Postgres --\nnobody could ever create an organisation there at all.",
        "operationId": "create_org_orgs_post",
        "parameters": [
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/{tenant_id}/invitations": {
      "post": {
        "tags": [
          "orgs"
        ],
        "summary": "Create Invitation",
        "operationId": "create_invitation_orgs__tenant_id__invitations_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invitations/accept": {
      "post": {
        "tags": [
          "orgs"
        ],
        "summary": "Accept Invitation",
        "description": "REQ-038: invited user reaches the intended project after sign-in.\n\nOrg bootstrap fix, two parts. First: the initial `Invitation` lookup is\nBY TOKEN, deliberately without knowing the tenant yet -- that is the\nwhole point of a self-contained invitation link. But `invitations` is\nRLS-protected the same way `memberships` is, and there is no tenant\ncontext to set beforehand here (unlike get_active_membership, the\ntenant isn't known until AFTER this exact query finds it). Migration\n0014 adds a second, narrowly-scoped permissive SELECT policy on\n`invitations` keyed on `app.invitation_lookup_token_hash` -- Postgres\ncombines multiple permissive policies for the same command with OR, so\na caller who sets this to the token hash they're presenting can find\nthat ONE row regardless of tenant context, without weakening the\ntenant-isolation policy for anything else (INSERT/UPDATE/DELETE, or a\nSELECT that doesn't set this). Knowing the unguessable, high-entropy\ntoken IS the authorization to read that row -- the same trust\nassumption an emailed invitation link already relies on. Second: once\nthe invitation's real tenant_id is known, context is set from it before\nthe `existing`-membership check and the INSERT below -- the same\nordering bug create_org had, for the same reason (no prior\nget_active_membership dependency to have set it already).",
        "operationId": "accept_invitation_invitations_accept_post",
        "parameters": [
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/{tenant_id}/me": {
      "get": {
        "tags": [
          "orgs"
        ],
        "summary": "My Membership",
        "description": "REQ-003: active tenant resolved from session and re-verified here,\nnever trusted from a client-supplied claim.",
        "operationId": "my_membership_orgs__tenant_id__me_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/{tenant_id}/approval-gate-check": {
      "get": {
        "tags": [
          "orgs"
        ],
        "summary": "Approval Gate Check",
        "description": "REQ-002 enforcement placeholder: proves the MFA gate for\napproval-adjacent roles works before WP07's actual decision endpoints\nexist to protect. Delete once POST /api/occurrences/{id}/decisions\n(Blueprint Sec.5.3/5.4) is built and carries this same dependency.",
        "operationId": "approval_gate_check_orgs__tenant_id__approval_gate_check_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Approval Gate Check Orgs  Tenant Id  Approval Gate Check Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/{tenant_id}/access-review": {
      "get": {
        "tags": [
          "orgs"
        ],
        "summary": "Access Review",
        "description": "WP12/REQ-047: 'review access quarterly' -- this is the mechanism\n(a complete, current membership listing for a tenant administrator to\nactually review) not the cadence itself, which stays an operational\npractice this endpoint doesn't schedule (same honest split as WP08's\ncheckpoint-frequency note: the capability exists, running it on a\nreal quarterly cycle is a process decision, not code). Includes\ninactive (revoked) memberships too -- a review that only shows who\ncurrently has access can't confirm past revocations actually took\neffect.",
        "operationId": "access_review_orgs__tenant_id__access_review_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccessReviewEntryOut"
                  },
                  "title": "Response Access Review Orgs  Tenant Id  Access Review Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInvitationRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "AcceptInvitationRequest"
      },
      "AccessReviewEntryOut": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "member_since": {
            "type": "string",
            "format": "date-time",
            "title": "Member Since"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "role",
          "active",
          "member_since"
        ],
        "title": "AccessReviewEntryOut"
      },
...
      "CreateOrgRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateOrgRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InviteOut": {
        "properties": {
          "invitation_id": {
            "type": "string",
            "title": "Invitation Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "invitation_id",
          "token"
        ],
        "title": "InviteOut"
      },
      "InviteRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "email",
          "role"
        ],
        "title": "InviteRequest"
      },
...
      "MembershipOut": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "role"
        ],
        "title": "MembershipOut"
      },
...
      "OrgOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "OrgOut"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "contributor",
          "approver",
          "sponsor",
          "assurance_reviewer",
          "tenant_administrator"
        ],
        "title": "Role",
        "description": "Blueprint Sec.2 role table. platform_operator is out of tenant scope\n(no membership row) and is not modelled here yet -- WP12."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}